SO107. Split CRS metadata into crs_definitions.py and use key-level hint from crs.json.
